Cleanups

+++ esajpip/esajpip/src/net/socket.cc
@@ -47,7 +47,7 @@ namespace net {
             return recv(sid, buf, len, 0);
 
         PollFD poll_fd(sid, POLLIN);
-        if (poll(&poll_fd, 1, 0) <= 0) return (ssize_t) -1;
+        if (poll(&poll_fd, 1, 0) <= 0) return -1;
         return recv(sid, buf, len, 0);
     }
 
@@ -58,7 +58,7 @@ namespace net {
             return recvfrom(sid, buf, len, 0, address->GetSockAddr(), &sock_len);
 
         PollFD poll_fd(sid, POLLIN);
-        if (poll(&poll_fd, 1, 0) <= 0) return (ssize_t) -1;
+        if (poll(&poll_fd, 1, 0) <= 0) return -1;
         return recvfrom(sid, buf, len, 0, address->GetSockAddr(), &sock_len);
     }
 
@@ -67,7 +67,7 @@ namespace net {
             return send(sid, buf, len, 0);
 
         PollFD poll_fd(sid, POLLOUT);
-        if (poll(&poll_fd, 1, 0) <= 0) return (ssize_t) -1;
+        if (poll(&poll_fd, 1, 0) <= 0) return -1;
         return send(sid, buf, len, 0);
     }
 
@@ -76,7 +76,7 @@ namespace net {
             return sendto(sid, buf, len, 0, address.GetSockAddr(), address.GetSize());
 
         PollFD poll_fd(sid, POLLOUT);
-        if (poll(&poll_fd, 1, 0) <= 0) return (ssize_t) -1;
+        if (poll(&poll_fd, 1, 0) <= 0) return -1;
         return sendto(sid, buf, len, 0, address.GetSockAddr(), address.GetSize());
     }
